+++ tests/conftest.py
@@ -7,25 +7,22 @@ import pytest
 @pytest.fixture
 def test_file(tmpdir):
     """Fixture for providing a test file to pass to mwcp."""
-    file_path = os.path.join(str(tmpdir), 'test.txt')
-    with open(file_path, 'wb') as f:
-        f.write(b"This is some test data!")
+    file_path = tmpdir / 'test.txt'
+    file_path = file_path.write_binary(b'This is some test data!')
     return file_path
 
 
 @pytest.fixture
 def test_dir(tmpdir):
     """Fixture for providing a test directory to pass to mwcp."""
-    directory = os.path.join(str(tmpdir), 'test_dir')
-    os.makedirs(directory)
+    directory = tmpdir.mkdir('test_dir')
     for i in range(5):
-        file_path = os.path.join(directory, 'test_{}.txt'.format(i))
-        with open(file_path, 'wb') as f:
-            f.write(b"This is some test data!")
+        file_path = directory / 'test_{}.txt'.format(i)
+        file_path.write_binary(b"This is some test data!")
     return directory
 
 
-TEST_PARSER = '''
+TEST_PARSER = u'''
 from mwcp import Parser
 
 class Downloader(Parser):
@@ -37,8 +34,8 @@ class Implant(Parser):
     
 '''
 
-TEST_PARSER_CONFIG = '''
-test_parser:
+TEST_PARSER_CONFIG = u'''
+Sample:
     description: A test parser
     author: Mr. Tester
     parsers:
@@ -48,18 +45,16 @@ test_parser:
 
 
 @pytest.fixture
-def test_parser(tmpdir):
+def Sample_parser(tmpdir):
     """Creates and returns the file path to a test parser."""
-    file_path = os.path.join(str(tmpdir), 'test_parser.py')
-    with open(file_path, 'w') as f:
-        f.write(TEST_PARSER)
+    parser_file = tmpdir / 'Sample.py'
+    parser_file.write_text(TEST_PARSER, 'utf8')
 
     # Parser directories must have an __init__.py
-    with open(os.path.join(str(tmpdir), '__init__.py'), 'w') as _:
-        pass
-        
-    config_path = os.path.join(str(tmpdir), 'parser_config.yml')
-    with open(config_path, 'w') as f:
-        f.write(TEST_PARSER_CONFIG)
-        
-    return file_path, config_path
+    init = tmpdir / '__init__.py'
+    init.write_text(u'', 'utf8')
+
+    config_file = tmpdir / 'parser_config.yml'
+    config_file.write_text(TEST_PARSER_CONFIG, 'utf8')
+
+    return parser_file, config_file