+++ tests/test_issues.py
@@ -12,20 +12,20 @@ def test_csv_row_bug(script_runner, tmpdir, test_dir):
     Tests bug where first row is formatted different from other rows.
     Occurs when outputting csv and input file is a directory.
     """
-    cwd = str(tmpdir)
-    csv_path = os.path.join(cwd, 'csv_file.csv')
+    csv_file = tmpdir / 'csv_file.csv'
 
-    ret = script_runner.run('mwcp-tool', '-p', 'foo', '-c', csv_path, test_dir, cwd=cwd)
+    ret = script_runner.run(
+        'mwcp-tool', '-p', 'foo', '-c', str(csv_file), str(test_dir), cwd=str(tmpdir))
     print(ret.stdout)
     print(ret.stderr, file=sys.stderr)
-
     assert ret.success
-    assert os.path.exists(csv_path)
 
-    with open(csv_path, 'r') as fo:
+    assert csv_file.exists()
+
+    with csv_file.open('r') as fo:
         reader = csv.reader(fo)
         rows = list(reader)
-        assert len(rows) == len(os.listdir(test_dir)) + 1
+        assert len(rows) == len(test_dir.listdir()) + 1
         assert rows[0] == ['scan_date', 'inputfilename', 'outputfile.name',
                            'outputfile.description', 'outputfile.md5', 'address', 'debug', 'url']
         for i, row in enumerate(rows[1:]):
@@ -38,7 +38,6 @@ def test_csv_row_bug(script_runner, tmpdir, test_dir):
                 '127.0.0.1',
                 ('[+] File test_{0}.txt identified as Foo.\n'
                 '[+] size of inputfile is 23 bytes\n'
-                '[+] Output file: fb843efb2ffec987db12e72ca75c9ea2_fooconfigtest.txt\n'
                 '[+] operating on inputfile test_{0}.txt').format(i),
                 'http://127.0.0.1',
             ]